[frontend] Integrate the new winternitz circuit in hash-based signature aggregation
This PR:

1. Removes the old winternitz_ots module, replacing it with the new
optimized winternitz_ots_bounded module.
2. Integrates the new winternitz_ots circuit int he xmss and
xmss_aggregate circuits.
3. Integrates the hash-based signature circuit into the hashsign circuit
example and updates the snapshot
