+++ crates/frontend/src/circuits/hash_based_sig/mod.rs
@@ -4,9 +4,6 @@ pub mod codeword;
 pub mod hashing;
 pub mod merkle_tree;
 pub mod winternitz_ots;
-pub mod winternitz_ots_bounded;
 pub mod witness_utils;
-pub mod witness_utils_bounded;
 pub mod xmss;
 pub mod xmss_aggregate;
-pub mod xmss_bounded;
